Changes to be compatible with our AQ reg form
Main problem is that the library doesn't work well with a password input
that is already part of a bootstrap 'input group', since this library
wraps the password input field and the 'show password' icon in an input
group, resulting in a nested input group (and visual garbage). This is
documented here: https://github.com/wenzhixin/bootstrap-show-password/issues/25

This removes that wrapper, and also appends the icon to the end of the
input group div, so that it plays nicely with our bootstrap-validator
error div.

See https://github.com/hinge-health/basilisk/pull/509 for more info.